move workflows out of subfolders and rename DreamGenericWorkflow to DreamWorkflow
