+++ src/ess/dream/workflows.py
@@ -6,34 +6,116 @@ import itertools
 import sciline
 import scipp as sc
 import scippnexus as snx
+from scippneutron.metadata import Software
 
 from ess.powder import providers as powder_providers
 from ess.powder import with_pixel_mask_filenames
 from ess.powder.correction import RunNormalization, insert_run_normalization
 from ess.powder.types import (
     AccumulatedProtonCharge,
+    BunkerMonitor,
+    CaveMonitor,
     CaveMonitorPosition,  # Should this be a DREAM-only parameter?
     EmptyCanRun,
     KeepEvents,
     PixelMaskFilename,
     Position,
+    ReducerSoftware,
     SampleRun,
+    TimeOfFlightLookupTableFilename,
     TofMask,
     TwoThetaMask,
     VanadiumRun,
     WavelengthMask,
 )
+from ess.reduce.nexus.types import DetectorBankSizes, NeXusName
 from ess.reduce.parameter import parameter_mappers
+from ess.reduce.time_of_flight import GenericTofWorkflow
 from ess.reduce.workflow import register_workflow
 
-from ..io.cif import (
+from .beamline import InstrumentConfiguration
+from .io.cif import (
     CIFAuthors,
     prepare_reduced_empty_can_subtracted_tof_cif,
     prepare_reduced_tof_cif,
 )
-from ..io.geant4 import providers as geant4_providers
-from ..parameters import typical_outputs
-from .generic import DreamGenericWorkflow
+from .io.geant4 import providers as geant4_providers
+from .parameters import typical_outputs
+
+DETECTOR_BANK_SIZES = {
+    "endcap_backward_detector": {
+        "strip": 16,
+        "wire": 16,
+        "module": 11,
+        "segment": 28,
+        "counter": 2,
+    },
+    "endcap_forward_detector": {
+        "strip": 16,
+        "wire": 16,
+        "module": 5,
+        "segment": 28,
+        "counter": 2,
+    },
+    "mantle_detector": {
+        "wire": 32,
+        "module": 5,
+        "segment": 6,
+        "strip": 256,
+        "counter": 2,
+    },
+    "high_resolution_detector": {"strip": 32, "other": -1},
+    "sans_detector": {"strip": 32, "other": -1},
+}
+
+
+def _get_lookup_table_filename_from_configuration(
+    configuration: InstrumentConfiguration,
+) -> TimeOfFlightLookupTableFilename:
+    from .data import tof_lookup_table_high_flux
+
+    match configuration:
+        case InstrumentConfiguration.high_flux_BC215:
+            out = tof_lookup_table_high_flux(bc=215)
+        case InstrumentConfiguration.high_flux_BC240:
+            out = tof_lookup_table_high_flux(bc=240)
+        case InstrumentConfiguration.high_resolution:
+            raise NotImplementedError("High resolution configuration not yet supported")
+
+    return TimeOfFlightLookupTableFilename(out)
+
+
+def _collect_reducer_software() -> ReducerSoftware:
+    return ReducerSoftware(
+        [
+            # Software.from_package_metadata('essdiffraction'),
+            Software.from_package_metadata('scippneutron'),
+            Software.from_package_metadata('scipp'),
+        ]
+    )
+
+
+def DreamWorkflow() -> sciline.Pipeline:
+    """
+    Dream generic workflow with default parameters.
+    The workflow is based on the GenericTofWorkflow.
+    It can load data from a NeXus file recorded on the DREAM instrument, and can
+    compute time-of-flight for the neutron events.
+
+    It can be used as is, or as a base for more specific workflows, such as
+    ``DreamPowderWorkflow``.
+    """
+    wf = GenericTofWorkflow(
+        run_types=[SampleRun, VanadiumRun, EmptyCanRun],
+        monitor_types=[BunkerMonitor, CaveMonitor],
+    )
+    wf[DetectorBankSizes] = DETECTOR_BANK_SIZES
+    wf[NeXusName[BunkerMonitor]] = "monitor_bunker"
+    wf[NeXusName[CaveMonitor]] = "monitor_cave"
+    wf.insert(_get_lookup_table_filename_from_configuration)
+    wf[ReducerSoftware] = _collect_reducer_software()
+    return wf
+
 
 _cif_providers = (
     prepare_reduced_tof_cif,
@@ -69,7 +151,7 @@ def DreamPowderWorkflow(*, run_norm: RunNormalization) -> sciline.Pipeline:
     :
         A workflow object for DREAM.
     """
-    wf = DreamGenericWorkflow()
+    wf = DreamWorkflow()
     for provider in itertools.chain(powder_providers, _cif_providers):
         wf.insert(provider)
     insert_run_normalization(wf, run_norm)
@@ -93,7 +175,7 @@ def DreamGeant4Workflow(*, run_norm: RunNormalization) -> sciline.Pipeline:
     :
         A workflow object for DREAM.
     """
-    wf = DreamGenericWorkflow()
+    wf = DreamWorkflow()
     for provider in itertools.chain(geant4_providers, powder_providers, _cif_providers):
         wf.insert(provider)
     insert_run_normalization(wf, run_norm)
@@ -157,5 +239,6 @@ __all__ = [
     'DreamGeant4ProtonChargeWorkflow',
     'DreamGeant4Workflow',
     'DreamPowderWorkflow',
+    'DreamWorkflow',
     'default_parameters',
 ]
