--- src/ess/dream/workflows/__init__.py
@@ -1,23 +1,0 @@
-# SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
-"""
-Workflows for the DREAM instrument.
-"""
-
-from .generic import DreamGenericWorkflow
-from .powder import (
-    DreamGeant4MonitorHistogramWorkflow,
-    DreamGeant4MonitorIntegratedWorkflow,
-    DreamGeant4ProtonChargeWorkflow,
-    DreamGeant4Workflow,
-    DreamPowderWorkflow,
-)
-
-__all__ = [
-    'DreamGeant4MonitorHistogramWorkflow',
-    'DreamGeant4MonitorIntegratedWorkflow',
-    'DreamGeant4ProtonChargeWorkflow',
-    'DreamGeant4Workflow',
-    'DreamGenericWorkflow',
-    'DreamPowderWorkflow',
-]
